Refactor - Abstract orders
